authorization left to be done :)

+++ database/migrations/2019_06_23_133654_create_expences_table.php
@@ -16,7 +16,7 @@ class CreateExpencesTable extends Migration
         Schema::create('expences', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('expence_name');
-            $table->integer('expence_amount');
+            $table->float('expence_amount');
             $table->timestamps();
         });
     }
